Ejercicio04 Ricardo

+++ Ejercicio04/Ejercicio04/src/ejercicio04/Ejercicio04.java
@@ -5,8 +5,9 @@ public class Ejercicio04 {
         int var2 = 1;
         int var3 = 3;
         int var4 = 2;
-        int resultado1 = (var1/(var2 + var3))+ var4;
+        double resultado1 = (var1/(var2 + var3))+ var4;
         boolean resultadofinal = resultado1 >= var3;
+        //Hacemos la comparacion de las dos cantidades con un operador relacional y obtenemos como resultado un booleano.
         System.out.println(resultadofinal);
     }
 }
